Remove useless test
The testLMSimpleShapeNGauss1 was testing two different PSF fits to ensure that they
are "close". Since the definition of close and is ambiguous and there is no
measure as to how close the PSF's being compared should be, this test was not
really testing anything. Instead it was removed but similar tests that
ensure that the functions were run correctly remain.